# docs/swagger.yaml
basePath: /api
definitions:
...
  request.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
  response.BacklogItemSummary:
    properties:
      id:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  response.SprintSummaryFull:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  response.UserActivitiesResponse:
    properties:
      activities:
        items:
          $ref: '#/definitions/response.UserActivityResponse'
        type: array
      limit:
        type: integer
      total:
        type: integer
    type: object
  response.UserActivityResponse:
    properties:
      action:
        type: string
      comment:
        type: string
      field_changed:
        type: string
      id:
        type: string
      item:
        $ref: '#/definitions/response.BacklogItemSummary'
      new_value: {}
      old_value: {}
      sprint:
        $ref: '#/definitions/response.SprintSummaryFull'
      timestamp:
        type: string
      type:
        description: '"item" or "sprint"'
        type: string
    type: object
  response.UserResponse:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  utils.ErrorDetail:
    properties:
      code:
        type: string
      details:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/utils.ErrorDetail'
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@sprintbacklog.com
    name: API Support
  description: API for Sprint Backlog - A project management tool for agile teams
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Sprint Backlog API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Get all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.UserResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
  /users/{id}:
    get:
      description: Get a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
  /users/{id}/activities:
    get:
      description: Get a user's activity history (item and sprint actions)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: Maximum number of activities to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.UserActivitiesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user activities
      tags:
      - users
  /users/profile:
    put:
      consumes:
      - application/json
      description: Update the current user's profile information
      parameters:
      - description: Update profile request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using Bearer scheme. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"